+++ src/multiset.h
@@ -14,7 +14,7 @@
 
 struct multiset
 {
-  ptrdiff_t refs;
+  INT32 refs;
 #ifdef PIKE_SECURITY
   struct object *prot;
 #endif
